+++ ne_version.h
@@ -1,3 +1,3 @@
 ﻿#pragma once
-#define NE_PUBLISHED L"28.05.2026 09:45"
-#define NE_VERSION   L"2026.05.28.09"
+#define NE_PUBLISHED L"28.05.2026 10:50"
+#define NE_VERSION   L"2026.05.28.10"